hostname translation + `Socket2` initialization
This commit sets up the ground work for the Ring
Software. A thing to note is that the usage
of `RAW` sockets require elevate previlage, so
you need to run the program as root. This commit also sets
up the translation for hostname directly to DNS. It doesn't
use third-party library for DNS look-up, and simply translate
using std functions. Also, you are free to provide port number or
not while running the program. It's bearing fruit.

~@CroxxN