First Working Commit
Yeah! It finally works. "Works" is a far-fetcehd term because
although it sends and receive echo message, it's extremely
rudementary, and doesn't handle any error. In comming commits,
I'll try to add error handeling, and also colorize terminal output
for status of the echo request. The previous commit was hilarious,
by the way. Because the program was intercepting it's own message,
and in actuality no message was being sent. I knew it because
the message type code was still 8 and not 0. Turns out I wasn't
properly adding checksum, so the checksum buffer was just void.
But I learned that if the checksum field is void, the
server just ignores that echo request.

~@CroxxN