feat: verify + seq + TTL + time
This commit adds a bunch of new features, the most
significant of which is the `verify` functionality.
Every packet will now be verified by calculating
that packet's checksum. Also, the current packet
sequence number is displayed along side the packet
status message, along with Time-to-Live value, and
the time taken for packet to reach the destination.